First files
Thet `/arduino` directory contains arduino scripts for the stewart
platform. `center.ino` was for testing the motors, finding the center
and limits of their range. `stewart_platform.ino` is the working code
for low-level control of the motors and kinematics of the platform. The
`/platform_files` directory has the images for lasercutting. There are 3
platform/base images, in both pdf and eps format, and the last two were
for working with the servo arm. The lab computer has svg files on it. I
should grab those. The `/simulations` directory was made for prototyping
the controls. I may throw it out.